Create AuthManager and add Estuary nonce
